LayoutTests/platform/ios/fast/forms/select/base/option-internal-inset-outline-offset-expected.txt

Tests that -internal-inset outline-offset resolves correctly with various outline widths, styles, and zoom.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


FAIL defaultOffset == "-3px" || defaultOffset == "-2px" should be true. Was false.
FAIL actualOffset should be -2px. Was 0px.
FAIL actualOffset should be -10px. Was 0px.
FAIL zoomedOffset == "-3px" || zoomedOffset == "-2px" should be true. Was false.
FAIL actualOffset should be -5px. Was 0px.
PASS successfullyParsed is true
Some tests failed.

TEST COMPLETE
